requirements.txt
flask>=2.0.1
flask-sqlalchemy>=2.5.1
pymysql>=1.0.2
python-dotenv>=0.19.0
flask-cors>=3.0.10
pytest>=7.0.0
pytest-flask>=1.2.0